Enforce uniqueness of URN for active/inactive projects
We add a `EnforceUniquenessOfProjectUrn` db trigger to raise an
error and rollback "after insert" if that insert is attempting
to add a new project record where an existing active or inactive
project with the same URN exists.

We've been seeing some duplicate transfers and conversions being
created by our automated "handover" process between "prepare" and
"complete". Our "complete" API receives POSTs to:

- `/api/v1/projects/conversions` and
- `/api/v1/projects/conversions`

these are handled by:

- `Api::Conversions::CreateProjectService` and
- `Api::Transfers::CreateProjectService`

which both use the `UrnUniqueForApiValidator` to perform this
validation at the application level, but for a reliable
constraint we need to use a control at the database layer.